Detail edits and refactor
